KOGITO-6694 : for debugging purpose, broken code
