+++ kogito-quarkus-examples/kogito-travel-agency/extended/docker-compose/.env
@@ -1,0 +1,2 @@
+KOGITO_VERSION=latest
+DOCKER_GATEWAY_HOST=kubernetes.docker.internal
