+++ kogito-quarkus-examples/kogito-travel-agency/extended/travels/src/main/java/org/acme/travels/VisaApplication.java
@@ -25,8 +25,6 @@ public class VisaApplication {
     private String passportNumber;
     private String nationality;
 
-    private boolean approved;
-
     public VisaApplication() {
 
     }
@@ -97,19 +95,10 @@ public class VisaApplication {
         this.nationality = nationality;
     }
 
-    public boolean isApproved() {
-        return approved;
-    }
-
-    public void setApproved(boolean approved) {
-        this.approved = approved;
-    }
-
     @Override
     public String toString() {
         return "VisaApplication [firstName=" + firstName + ", lastName=" + lastName + ", city=" + city + ", country="
-                + country + ", duration=" + duration + ", passportNumber=" + passportNumber + ", approved=" + approved
-                + "]";
+                + country + ", duration=" + duration + ", passportNumber=" + passportNumber + "]";
     }
 
 }
