+++ kogito-quarkus-examples/kogito-travel-agency/extended/visas/src/main/resources/application.properties
@@ -11,6 +11,11 @@ kogito.service.url=http://localhost:8090
 kogito.dataindex.http.url=http://localhost:8180
 kogito.dataindex.ws.url=ws://localhost:8180
 
+quarkus.oidc.auth-server-url=http://localhost:8281/auth/realms/kogito
+quarkus.oidc.client-id=kogito-app
+quarkus.oidc.credentials.secret=secret
+
+
 quarkus.infinispan-client.server-list=localhost:11222
 quarkus.infinispan-client.use-auth=false
 #quarkus.infinispan-client.server-list=kogito-infinispan:11222
